rename utils folder to helpers
