--- packages/vue3/src/utils/type.util.ts
@@ -1,3 +1,0 @@
-export function launder<T>(arg: unknown) {
-  return arg as T;
-}
